refactor(dm05): rename stats contract helpers
