+++ src/lerobot/policies/dm05/modeling_dm05.py
@@ -36,7 +36,7 @@ from ..pretrained import ActionSelectKwargs, PreTrainedPolicy, T
 from .configuration_dm05 import DM05Config
 from .constants import ACTION_REFERENCE_OFFSET
 from .conversion_dm05 import DM05LerobotBatchConverter
-from .stats_dm05 import (
+from .stats_validation_dm05 import (
     dm05_prepare_stats_command,
     dm05_stats_complete,
     validate_dm05_relative_action_stats,
