+++ src/lerobot/policies/dm05/prepare_stats_dm05.py
@@ -33,7 +33,7 @@ from lerobot.datasets.utils import STATS_PATH, serialize_dict
 from lerobot.utils.constants import ACTION, OBS_STATE
 
 from .configuration_dm05 import DM05Config
-from .stats_dm05 import (
+from .stats_validation_dm05 import (
     dm05_feature_stats_complete,
     dm05_stats_complete,
     validate_dm05_relative_action_stats,
